Class 2: Assignment 1
Write a program where the user enters information on the prompt (product’s name, price and qty) and save the info in variables, then you have to display on the console the info, subtotal and the final price adding taxes. You should have:

- price Calculation
- tax Calculation
- display Information

Display the results in HTML like a receipt.